Extract superclass for custom date parsers
